Fix the SDK
Keep missing other than User entity implementation but stoppping there 
because of a potential change to Flutter/Dart

+++ SDK/Split/Split/Constants.cs
@@ -6,6 +6,6 @@
         public static readonly string APIBaseURI = BaseURI + "/api";
         public static readonly string AuthentificationEndpoint = APIBaseURI + "/login";
 
-        public static readonly int InvalidId = -1;
+        public static readonly long InvalidId = -1;
     }
 }
